move reducers to their own domain
